ttl correction for further usage
