+++ ydb/core/tx/columnshard/engines/scheme/tiering/common.cpp
@@ -1,0 +1,5 @@
+#include "common.h"
+
+namespace NKikimr::NOlap::NTiering::NCommon {
+
+}
