+++ ydb/core/tx/columnshard/engines/scheme/tiering/common.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <util/generic/string.h>
+
+namespace NKikimr::NOlap::NTiering::NCommon {
+
+static inline const TString DeleteTierName = "$$DELETE";
+
+}
