+++ ydb/core/tx/columnshard/engines/scheme/tiering/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    tier_info.cpp
+    common.cpp
+)
+
+PEERDIR(
+    ydb/core/formats/arrow/serializer
+)
+
+END()
