+++ ydb/core/tx/tiering/rule/manager.cpp
@@ -15,6 +15,9 @@ NMetadata::NModifications::TOperationParsingResult TTieringRulesManager::DoBuild
     TInternalModificationContext& /*context*/) const {
     NMetadata::NInternal::TTableRecord result;
     result.SetColumn(TTieringRule::TDecoder::TieringRuleId, NMetadata::NInternal::TYDBValue::Utf8(settings.GetObjectId()));
+    if (settings.GetObjectId().StartsWith("$")) {
+        return TConclusionStatus::Fail("tiering rule name cannot contains '$' character");
+    }
     {
         auto fValue = settings.GetFeaturesExtractor().Extract(TTieringRule::TDecoder::DefaultColumn);
         if (fValue) {
